+++ Jmol/src/org/openscience/jmol/render/VectorShape.java
@@ -4,7 +4,7 @@ import org.openscience.jmol.*;
 
 import java.awt.Graphics;
 import javax.vecmath.Point3f;
-import javax.vecmath.Matrix4d;
+import javax.vecmath.Matrix4f;
 
 public class VectorShape implements Shape, Transformable {
 
@@ -31,7 +31,7 @@ public class VectorShape implements Shape, Transformable {
     return buffer.toString();
   }
 
-  public void transform(Matrix4d matrix) {
+  public void transform(Matrix4f matrix) {
     matrix.transform(origPoint, screenPositionOrig);
     matrix.transform(endPoint, screenPositionEnd);
   }
